# Hypatiamat-BackOffice/apiDados/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Api Hypatiamat",
        "description": "Aqui encontra-se a documentação de toda a api de dados do Hypatiamat, incluindo todas as rotas e toda a informação à cerca do que é precisar passar como parâmetros, bem como os resultados que se podem esperar ao executar aquela rota."
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3050",
    "basePath": "/",
...
    "paths": {
...
        "/aplicacoes/professores": {
            "get": {
                "tags":[
                    "Professores"
                ],
                "summary": "Todos os professores registados.",
                "description": "Retorna uma lista com toda informação à cerca de todos os professores.",
                "responses": {
                    "200": {
                        "description": "Lista de Professores",
                        "schema": {
                          "type":"array",
                          "items":{
                            "$ref": "#/definitions/Professor"
                          }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            },
            "post":{
                "tags":[
                    "Professores"
                ],
                "summary": "Inserir um novo professor.",
                "description": "Adiciona um novo curso à base de dados.",
                "parameters": [
                    {
                      "name": "codigo",
                      "in": "formData",
                      "description": "Código/Username de um professor",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "nome",
                      "in": "formData",
                      "description": "Nome do professor",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "escola",
                      "in": "formData",
                      "description": "Identificador da escola do professor",
                      "required": true,
                      "type": "integer"
                    },
                    {
                      "name": "email",
                      "in": "formData",
                      "description": "Email do professor",
                      "required": true,
                      "type": "string"
                    },{
                      "name": "password",
                      "in": "formData",
                      "description": "Password do professor",
                      "required": true,
                      "type": "string",
                      "format": "password"
                    },{
                      "name": "confirmacao",
                      "in": "formData",
                      "description": "0 se o professor não está confirmado, 1 caso contrário",
                      "required": true,
                      "type": "integer"
                    },{
                      
                      "name": "premium",
                      "in": "formData",
                      "description": "Nível de acesso",
                      "required": true,
                      "type": "integer"
                    },{
                      "name": "socionum",
                      "in": "formData",
                      "description": "Número de sócio do professor",
                      "required": true,
                      "type": "integer"
                    },{
                      "name": "validade",
                      "in": "formData",
                      "description": "Data em que a sua conta expira",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "projeto",
                      "in": "formData",
                      "description": "Projeto",
                      "required": true,
                      "type": "integer"
                    }

                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id do professor criado",
                        "schema": {
                            "type": "string",
                            "items": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/aplicacoes/professores/{codigoProf}": {
            "get": {
              "tags": [
                "Professores"
              ],
              "summary": "Dado um código de um professor, devolve toda a informação relativa a esse professor.",
              "description": "Devolve toda a informação relativa a um professor.",
              "operationId": "getProfessorByCodigo",
              "parameters": [
                {
                  "name": "codigoProf",
                  "in": "path",
                  "description": "Código do professor a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "Retorna a informação relativa a um professor.",
                    "schema": {
                        "$ref": "#/definitions/Professor"   
                        }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                },
                "403":{
                  "description": "Não possuí nível de acesso suficiente."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            },
            "put":{
              "tags": [
                "Professores"
              ],
              "summary": "Dado um código de um professor, edite a informação relacionada com ele.",
              "description": "Editar um professor.",
              "operationId": "putProfessorByCodigo",
              "parameters": [
                {
                  "name": "codigoProf",
                  "in": "path",
                  "description": "Código/Username de um professor",
                  "required": true,
                  "type": "string"
                },
                {
                  "name": "nome",
                  "in": "formData",
                  "description": "Nome do professor",
                  "required": true,
                  "type": "string"
                },
                {
                  "name": "escola",
                  "in": "formData",
                  "description": "Identificador da escola do professor",
                  "required": true,
                  "type": "integer"
                },
                {
                  "name": "email",
                  "in": "formData",
                  "description": "Email do professor",
                  "required": true,
                  "type": "string"
                },{
                  "name": "password",
                  "in": "formData",
                  "description": "Password do professor",
                  "required": true,
                  "type": "string"
                },{
                  "name": "confirmacao",
                  "in": "formData",
                  "description": "0 se o professor não está confirmado, 1 caso contrário",
                  "required": true,
                  "type": "integer"
                },{
                  
                  "name": "premium",
                  "in": "formData",
                  "description": "1 se o professor é premium, 0 caso contrário",
                  "required": true,
                  "type": "string"
                },{
                  "name": "socionum",
                  "in": "formData",
                  "description": "Número de sócio do professor",
                  "required": true,
                  "type": "integer"
                },{
                  "name": "validade",
                  "in": "formData",
                  "description": "Data em que a sua conta expira",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                  "description": "O professor foi editado."
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                },
                "403":{
                  "description": "Não possuí nível de acesso suficiente."
                }
              },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            },
            "delete": {
              "tags": [
                "Professores"
              ],
              "summary": "Dado um código de professor, elimine o professor da base de dados.",
              "description": "Apagar um professor.",
              "operationId": "deleteProfessorByCodigo",
              "parameters": [
                {
                  "name": "codigoProf",
                  "in": "path",
                  "description": "Código do professor a apagar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                  "description": "O professor foi eliminado."
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                },
                "403":{
                  "description": "Não possuí nível de acesso suficiente."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
              
            }
        },
        "/aplicacoes/professores/{idProf}/turmas": {
            "get": {
              "tags": [
                "Professores"
              ],
              "summary": "Dado um identificador de um professor, devolve todas as suas turmas.",
              "description": "Devolve todos as turmas de um professor.",
              "operationId": "getTurmasProfessorByCodigo",
              "parameters": [
                {
                  "name": "idProf",
                  "in": "path",
                  "description": "Id do professor a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                      "description": "Lista de Turmas",
                      "schema": {
                        "type":"array",
                        "items":{
                          "$ref": "#/definitions/Turma"
                        }
                      }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            }
        },
        "/aplicacoes/professores/{codigoProf}/password": {
            "put": {
              "tags": [
                "Professores"
              ],
              "summary": "Dado um código de um professor, edita-se a sua palavra passe.",
              "description": "Editar a password de um professor.",
              "operationId": "putPasswordProfessorByCodigo",
              "parameters": [
                {
                  "name": "codigoProf",
                  "in": "path",
                  "description": "Código do professor a editar.",
                  "required": true,
                  "type": "string"
                },
                {
                  "name": "password",
                  "in": "formData",
                  "description": "Nova palavra passe.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "Palavra passe editada com sucesso."
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            }
          },
...
      "/aplicacoes/escolas/{idEscola}/professores":{
        "get": {
          "tags": [
            "Escolas"
          ],
          "summary": "Dado um identifcador da escola, devolve todos os professores da escola.",
          "description": "Devolve todos os professores de uma escola.",
          "operationId": "getProfessoresByEscola",
          "parameters": [
            {
              "name": "idEscola",
              "in": "path",
              "description": "Identificador da escola.",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
                "description": "Retorna todos os professores da escola.",
                "schema": {
                    "type":"array",
                    "items": {
                      "$ref": "#/definitions/Professor"
                    }   
                    }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      },
...
    },
    "securityDefinitions":{
      "api_key":{
        "type": "apiKey",
        "in": "query",
        "name": "token"
      }
    },
    "security": [{
       "api_key": []
    }],
    "definitions": {
...
        "Professor":{
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "codigo": {
              "type": "string"
            },
            "nome": {
              "type": "string"
            },
            "escola": {
              "type": "string"
            },
            "email":{
              "type":"string"
            },
            "confirmacao":{
              "type":"integer"
            },
            "premium":{
              "type":"integer"
            },
            "validade":{
              "type":"string"
            },
            "socionum":{
              "type":"integer"
            },
            "projeto":{
              "type":"integer"
            }
          }    
        },
        "Turma":{
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "idprofessor": {
              "type": "string"
            },
            "turma":{
              "type":"string"
            }
          }
        },
...
    }   
}